Update Interactor and new ViewModelInteractor to support @StateObject
See `Interactor` documentation for details. In summary, this addresses the issue where duplicate interactor instances might be retained as a result of SwiftUI updating a view.

The view must declare and reference the interactor as a `@StateObject`. This allows different instances of the view to reference the same interactor instance, therefore maintaining the states of the view's data. This occurs when multiple instances of the same view are instantiated over time by SwiftUI, as the view updates and changes due to data changes or user interactions. And because the interactor is a `@StateObject`, the binding of the interactor via the `bind(observer:)` modifier must be inside the view's `body`.